added plan model and ability to create the default plan on the fly, if one does not exist.
added terms of service check on signup model.

marking signup record as completed (via completed_at) instead of
destroying during customer creation. also, associating signup record
with the customer.

cleanup the signup subscription javascript to assign payment nonce and
type to signup variables in the form, instead of relying on Braintree's
default behaviour for the Dropin UI.

capturing the payment type, "CreditCard" or "PayPal", but be nice to
have data.